requirements.txt
azure-functions==1.21.3
pillow==11.0.0

keras==2.1.2
tensorflow==1.4.1
scipy==1.0.0
pyclipper==1.0.6
numpy==1.13.3
matplotlib==2.1.1
h5py==2.7.1
tables==3.4.2
